CookiesConfigSetting moved out from ui subpackage
